# Optimize conversion of MouseAction enum from/to string, remove allocations